estate_cms/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.7.2
    # via django
build==1.0.3
    # via pip-tools
chardet==5.2.0
    # via reportlab
click==8.1.7
    # via pip-tools
cssselect2==0.7.0
    # via svglib
django==5.0.2
    # via
    #   django-appconf
    #   django-classy-tags
    #   django-cms
    #   django-entangled
    #   django-filer
    #   django-fontawesome-5
    #   django-formtools
    #   django-parler
    #   django-polymorphic
    #   django-sekizai
    #   django-select2
    #   django-treebeard
    #   djangocms-admin-style
    #   djangocms-alias
    #   djangocms-frontend
    #   djangocms-versioning
    #   easy-thumbnails
django-appconf==1.0.6
    # via django-select2
django-classy-tags==4.1.0
    # via
    #   django-cms
    #   django-sekizai
django-cms==4.1.0
    # via
    #   djangocms-alias
    #   djangocms-attributes-field
    #   djangocms-frontend
    #   djangocms-text-ckeditor
    #   djangocms-versioning
django-entangled==0.5.4
    # via djangocms-frontend
django-filer==3.1.1
    # via djangocms-frontend
django-fontawesome-5==1.0.18
    # via -r requirements.in
django-formtools==2.5.1
    # via django-cms
django-fsm==2.8.1
    # via djangocms-versioning
django-ordered-model==3.7.4
    # via -r requirements.in
django-parler==2.3
    # via
    #   djangocms-alias
    #   djangocms-frontend
django-polymorphic==3.1.0
    # via django-filer
django-sekizai==4.1.0
    # via django-cms
django-select2==8.1.2
    # via djangocms-frontend
django-treebeard==4.7.1
    # via django-cms
djangocms-admin-style==3.3.0
    # via django-cms
djangocms-alias==2.0.0
    # via
    #   -r requirements.in
    #   djangocms-frontend
djangocms-attributes-field==3.0.0
    # via djangocms-frontend
djangocms-frontend[cms-4]==1.2.2
    # via -r requirements.in
djangocms-text-ckeditor==5.1.5
    # via djangocms-frontend
djangocms-versioning==2.0.0
    # via
    #   -r requirements.in
    #   djangocms-frontend
easy-thumbnails[svg]==2.8.5
    # via
    #   django-filer
    #   djangocms-frontend
html5lib==1.1
    # via djangocms-text-ckeditor
lxml==5.1.0
    # via svglib
packaging==23.2
    # via
    #   build
    #   django-cms
    #   djangocms-text-ckeditor
pillow==10.2.0
    # via
    #   djangocms-text-ckeditor
    #   easy-thumbnails
    #   reportlab
pip-tools==7.3.0
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
reportlab==4.1.0
    # via
    #   easy-thumbnails
    #   svglib
six==1.16.0
    # via html5lib
sqlparse==0.4.4
    # via django
svglib==1.5.1
    # via easy-thumbnails
tinycss2==1.2.1
    # via
    #   cssselect2
    #   svglib
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
wheel==0.42.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
